*/*: remove OpenMP detection and $OPENMP_CFLAGS
OpenMP is not used in m4rie, and its flags are not needed to build
m4rie against an m4ri that _does_ have OpenMP support. The one
scenario where m4ri/m4rie might need to agree on the matter is when
linking a static executable that uses them both; but for this, it will
be easier to rely on pkg-config to provide the correct linker flags
(and only when necessary).